+++ CHANGELOG.md
@@ -50,6 +50,7 @@ _The file [UPGRADING.md](UPGRADING.md) contains important information about upgr
 - core: fix buffer selected when searching by partial name and another buffer has a similar name ([#700](https://github.com/weechat/weechat/issues/700))
 - irc: fix memory leak in SASL authentication with mechanism "ecdsa-nist256p-challenge"
 - javascript, lua, perl, php, ruby, tcl: fix return value of function hdata_longlong
+- logger: add support of extra date/time specifiers like "%@" (UTC date) in options logger.file.mask, logger.file.path and logger.mask.* ([#2361](https://github.com/weechat/weechat/issues/2361))
 - relay/irc: remove message tags not negotiated by the client in redirected messages ([#2359](https://github.com/weechat/weechat/issues/2359))
 - relay/api: fix notify tag in line when already read on remote ([#2343](https://github.com/weechat/weechat/issues/2343))